The routing process for Cafe and Review was externalized to each router file.
