+++ schema.js
@@ -3,7 +3,7 @@ const Joi = require('joi');
 
 // cafeスキーマを定義
 module.exports.cafeSchema = Joi.object({
-    cafeData: Joi.object({ // リクエストボディにcafeDataがあることを定義
+    cafe: Joi.object({ // リクエストボディにcafeDataがあることを定義
         //　cafeDataの中身がどういう値かをobjectの中に定義する
         shopName: Joi.string().required(), // String&必須項目
         priceOfcoffee: Joi.number().min(0).required(), // 数値型&必須項目&最小値は0
@@ -13,7 +13,7 @@ module.exports.cafeSchema = Joi.object({
         confort: Joi.number().min(0).max(10).required(), // 数値&必須項目&最小値０&最大値10
         comment: Joi.string(), // String
         registerDate: Joi.string(), // String
-        image: Joi.string(), // String
+        image: Joi.string().allow(''), // String
     }).required() // リクエストボディのcafeDataは必須項目であることを定義
 });
