+++ views/cafes/caferegister.ejs
@@ -33,8 +33,8 @@
                             min="1" max="100">
                     </div>
                     <div class="form-group">
-                        <label for="comfort">快適さ</label>
-                        <input type="range" class="form-control-range" id="comfort" name="cafe[comfort]" min="1"
+                        <label for="confort">快適さ</label>
+                        <input type="range" class="form-control-range" id="confort" name="cafe[confort]" min="1"
                             max="10">
                     </div>
                     <div class="form-group">
